requirements.txt
certifi==2022.12.7
chardet==3.0.4
click==6.7
grab-screen==0.0.13
idna==2.6
imutils==0.5.4
keyboard==0.13.5
MouseInfo==0.1.3
mss==3.0.1
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.5.27
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.18.4
urllib3==1.22
